throwing datatable into its own component
